Até a atividade 10

+++ 01.py
@@ -6,8 +6,8 @@ def limpar():
 
 limpar()
 print(" - Quem é o maior? -")
-n1 = int(input(" Digite um número: \n  > "))
-n2 = int(input(" Digite outro número: \n  > "))
+n1 = float(input(" Digite um número: \n  > "))
+n2 = float(input(" Digite outro número: \n  > "))
 if n1 > n2:
 	print(f"O primeiro número é o maior. O primeiro número foi: {n1}")
 elif n2 > n1:
